Implementing user roles

+++ app.py
@@ -2,11 +2,11 @@ from flask import Flask, request, jsonify
 from models.user import User
 from database import db
 from flask_login import LoginManager, current_user,  login_user, logout_user, login_required
-
+import bcrypt
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = 'your_secret_key'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db' 
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://admin:admin123@127.0.0.1:3306/flask-crud' 
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -24,7 +24,7 @@ def login():
     username = data.get('username')
     password = data.get('password')
    
-    if username and password:
+    if username and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
         user = User.query.filter_by(username=username).first()  
         if user and user.password == password:
             login_user(user)
@@ -36,7 +36,7 @@ def login():
 @login_required
 def logout():
     logout_user()
-    return jsonify({"message: Logout realixado com sucesso!"})       
+    return jsonify({"message": "Logout realizado com sucesso!"})       
 
 @app.route('/user', methods=["POST"])
 def create_user():
@@ -46,7 +46,8 @@ def create_user():
     password = data.get('password')
 
     if username and email and password:
-        user = User(username=username, email=email, password=password)
+        hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+        user = User(username=username, email=email, password=hashed_password, role="user")
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "Usuario com registado com sucesso"}), 200
@@ -67,9 +68,11 @@ def update_users(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role== 'user':
+        return jsonify({"message": "Usuário não permitida"}), 403
     if user and data.get("password"):
         user.password = data.get("password")
-        db.commit()
+        db.session.commit()
         return jsonify({"message": f"Usuário {id_user} actualizado com sucesso"})
     return jsonify({"message": "usuário não encontrado"}), 404  
 
@@ -77,14 +80,14 @@ def update_users(id_user):
 @login_required
 def delete_users(id_user):
     user = User.query.get(id_user)
-
+    if current_user.role != 'admin':
+        return jsonify({"message": "Operação não permitida"}), 403 
     if id_user != current_user.id:
         return jsonify({"message": "Não permite apagar"}), 403  
-
     if user:
         db.session.delete(user)
-        db.session.commit
-        return jsonify({"message": f"Usuário {id_user} actualizado com sucesso "})
+        db.session.commit()
+        return jsonify({"message": f"Usuário {id_user} apagado com sucesso"})
     return jsonify({"message": "usuário não encontrado"}), 404  
 
 
@@ -93,6 +96,4 @@ def index():
     return "Hello, World!"
 
 if __name__ == '__main__':
-    with app.app_context():
-        db.create_all()
     app.run(debug=True)
